requirements.txt
python-binance==0.7.9
python-dotenv==0.15.0
python-telegram-bot==13.1
